+++ oldlx/Trees/Model.java
@@ -18,7 +18,7 @@ import heronarts.lx.model.LXPoint;
 import heronarts.lx.transform.LXTransform;
 
 
-class EntwinedBranch{
+class EntwinedBranch {
     /**
      * This defines the available mounting points on a given branch variation. The variable names and ratios for the keypoints
      * reflect what is in the CAD drawings for the branches
@@ -33,8 +33,8 @@ class EntwinedBranch{
     EntwinedBranch(int canopyMajorLength, int rotationalPosition, int layerBaseHeight) {
         int rotationIndex = rotationalPosition > 4 ? 4 - rotationalPosition % 4 : rotationalPosition;
         float canopyScaling = canopyMajorLength / 180;
-        double branchLengthRatios[] = { 0.37, 0.41, 0.50, 0.56, 0.63 };
-        double heightAdjustmentFactors[] = { 1.0, 0.96, 0.92, 0.88, 0.85 };
+        double branchLengthRatios[] = {0.37, 0.41, 0.50, 0.56, 0.63};
+        double heightAdjustmentFactors[] = {1.0, 0.96, 0.92, 0.88, 0.85};
         double branchLength = canopyMajorLength * branchLengthRatios[rotationIndex];
         xKeyPoints[4] = branchLength;
         xKeyPoints[3] = branchLength * 0.917;
@@ -81,7 +81,7 @@ class EntwinedBranch{
 
 }
 
-class EntwinedLayer{
+class EntwinedLayer {
     List<EntwinedBranch> branches;
 
     EntwinedLayer(int canopyMajorLength, int layerType, int layerBaseHeight) {
@@ -89,16 +89,16 @@ class EntwinedLayer{
         int rotationalPositions[];
         switch (layerType) {
             case 0:
-                rotationalPositions = new int[] { 0, 1, 2, 3, 4, 5, 6, 7 };
+                rotationalPositions = new int[]{0, 1, 2, 3, 4, 5, 6, 7};
                 break;
             case 1:
-                rotationalPositions = new int[] { 0, 2, 4, 6 };
+                rotationalPositions = new int[]{0, 2, 4, 6};
                 break;
             case 2:
-                rotationalPositions = new int[] { 1, 3, 5, 7 };
+                rotationalPositions = new int[]{1, 3, 5, 7};
                 break;
             default:
-                rotationalPositions = new int[] {};
+                rotationalPositions = new int[]{};
         }
         for (int i = 0; i < rotationalPositions.length; i++) {
             EntwinedBranch b = new EntwinedBranch(canopyMajorLength, rotationalPositions[i], layerBaseHeight);
@@ -107,8 +107,6 @@ class EntwinedLayer{
         this.branches = Collections.unmodifiableList(_branches);
     }
 }
-
-
 
 
 class Model extends LXModel {
@@ -122,16 +120,23 @@ class Model extends LXModel {
      * Cubes in the model
      */
     public final List<Cube> cubes;
+
+
+    public final List<BaseCube> baseCubes;
+
     public final Map<String, Cube[]> ipMap = new HashMap();
 
     private final ArrayList<ModelTransform> modelTransforms = new ArrayList<ModelTransform>();
     private final List<TreeConfig> treeConfigs;
 
     Model(List<TreeConfig> treeConfigs, List<CubeConfig> cubeConfig, List<ShrubConfig> shrubConfigs,
-            List<ShrubCubeConfig> shrubCubeConfig) {
+          List<ShrubCubeConfig> shrubCubeConfig) {
+
         super(new Fixture(treeConfigs, cubeConfig, shrubConfigs, shrubCubeConfig));
-        this.treeConfigs = treeConfigs;
+
         Fixture f = (Fixture) this.fixtures.get(0);
+
+        this.treeConfigs = treeConfigs;
         List<Cube> _cubes = new ArrayList<Cube>();
         List<CubeConfig> _inactiveCubeConfigs = new ArrayList();
         this.trees = Collections.unmodifiableList(f.trees);
@@ -149,6 +154,22 @@ class Model extends LXModel {
             _shrubCubes.addAll(shrub.cubes);
         }
         this.shrubCubes = Collections.unmodifiableList(_shrubCubes);
+
+
+        // Adding all cubes to baseCubes
+        List<BaseCube> _baseCubes = new ArrayList<BaseCube>();
+
+        for (Tree tree : this.trees) {
+            //ipMap.putAll(tree.ipMap);
+            _baseCubes.addAll(tree.cubes);
+        }
+
+        for (Shrub shrub : this.shrubs) {
+            //shrubIpMap.putAll(shrub.ipMap);
+            _baseCubes.addAll(shrub.cubes);
+        }
+        this.baseCubes = Collections.unmodifiableList(_baseCubes);
+
     }
 
     private static class Fixture extends LXAbstractFixture {
@@ -158,7 +179,7 @@ class Model extends LXModel {
         final List<Shrub> shrubs = new ArrayList<Shrub>();
 
         private Fixture(List<TreeConfig> treeConfigs, List<CubeConfig> cubeConfigs, List<ShrubConfig> shrubConfigs,
-                List<ShrubCubeConfig> shrubCubeConfigs) {
+                        List<ShrubCubeConfig> shrubCubeConfigs) {
             for (int i = 0; i < treeConfigs.size(); i++) {
                 TreeConfig tc = treeConfigs.get(i);
                 trees.add(new Tree(cubeConfigs, i, tc.x, tc.z, tc.ry, tc.canopyMajorLengths, tc.layerBaseHeights));
@@ -214,7 +235,7 @@ class Model extends LXModel {
         for (Cube cube : cubes) {
             cube.didTransform();
         }
-        
+
         for (ShrubCube cube : shrubCubes) {
             cube.resetTransform();
         }
@@ -244,7 +265,6 @@ class Model extends LXModel {
     private final List<ShrubConfig> shrubConfigs;
 
 
-
     public Vec3D getShrubMountPoint(ShrubCubeConfig c) {
         Vec3D p = null;
         Shrub shrub;
@@ -259,10 +279,9 @@ class Model extends LXModel {
         }
     }
 
-//    public void addShrubModelTransform(ShrubModelTransform modelTransform) {
-//        shrubModelTransforms.add(modelTransform);
-//    }
-
+    //    public void addShrubModelTransform(ShrubModelTransform modelTransform) {
+    //        shrubModelTransforms.add(modelTransform);
+    //    }
     public void runShrubTransforms() {
         for (ShrubCube cube : shrubCubes) {
             cube.resetTransform();
@@ -281,20 +300,22 @@ class Model extends LXModel {
     public void addModelTransform(Effect shrubModelTransform) {
         shrubModelTransforms.add((ModelTransform) shrubModelTransform);
 
-        
+
     }
 }
 
-class CubeConfig {
-    int treeIndex;
-    int layerIndex;
-    int branchIndex;
-    int mountPointIndex;
-    String ipAddress;
-    int outputIndex;
-    int cubeSizeIndex;
-    boolean isActive;
-}
+// class CubeConfig {
+//     int treeIndex;
+//     int layerIndex;
+//     int branchIndex;
+//     int mountPointIndex;
+//     String ipAddress;
+//     int outputIndex;
+//     int cubeSizeIndex;
+//     boolean isActive;
+// }
+//
+
 class TreeConfig {
     float x;
     float z;
@@ -399,7 +420,7 @@ class Tree extends LXModel {
                 Cube[] ndbCubes = entry.getValue();
                 for (int i = 0; i < 16; i++) {
                     if (ndbCubes[i] == null) { // fill all empty outputs with an inactive cube. Maybe this would be nicer to do at
-                                               // the model level in the future.
+                        // the model level in the future.
                         CubeConfig cc = new CubeConfig();
                         cc.treeIndex = treeIndex;
                         cc.branchIndex = 0;
@@ -432,136 +453,25 @@ class Tree extends LXModel {
     }
 }
 
-class Cube extends LXModel {
+class Cube extends BaseCube {
+  public static final int[] PIXELS_PER_CUBE = { 6, 6, 6, 12, 12 }; // Tiny cubes actually have less, but for Entwined we want to
+                                                                   // tell the NDB that everything is 6
+  public static final float[] CUBE_SIZES = { 4f, 7.5f, 11.25f, 15f, 16.5f };
+  /**
+   * Size of this cube, one of SMALL/MEDIUM/LARGE/GIANT
+   */
+  public final float size;
 
-    public static final int[] PIXELS_PER_CUBE = { 6, 6, 6, 12, 12 }; // Tiny cubes actually have less, but for Entwined we want to
-                                                                     // tell the NDB that everything is 6
-    public static final float[] CUBE_SIZES = { 4f, 7.5f, 11.25f, 15f, 16.5f };
-
-    /**
-     * Index of this cube in color buffer, colors[cube.index]
-     */
-    public final int index;
-
-    /**
-     * Size of this cube, one of SMALL/MEDIUM/LARGE/GIANT
-     */
-    public final float size;
-
-    public final int pixels;
-
-    /**
-     * Global x-position of center of cube
-     */
-    public final float x;
-
-    /**
-     * Global y-position of center of cube
-     */
-    public final float y;
-
-    /**
-     * Global z-position of center of cube
-     */
-    public final float z;
-
-    /**
-     * Pitch of cube, in degrees, relative to cluster
-     */
-    public final float rx;
-
-    /**
-     * Yaw of cube, in degrees, relative to cluster, after pitch
-     */
-    public final float ry;
-
-    /**
-     * Roll of cube, in degrees, relative to cluster, after pitch+yaw
-     */
-    public final float rz;
-
-    /**
-     * Local x-position of cube, relative to cluster
-     */
-    public final float lx;
-
-    /**
-     * Local y-position of cube, relative to cluster
-     */
-    public final float ly;
-
-    /**
-     * Local z-position of cube, relative to cluster
-     */
-    public final float lz;
-
-    /**
-     * x-position of cube, relative to center of tree base
-     */
-    public final float tx;
-
-    /**
-     * y-position of cube, relative to center of tree base
-     */
-    public final float ty;
-
-    /**
-     * z-position of cube, relative to center of tree base
-     */
-    public final float tz;
-
-    /**
-     * Radial distance from cube center to center of tree in x-z plane
-     */
-    public final float r;
-
-    /**
-     * Angle in degrees from cube center to center of tree in x-z plane
-     */
-    public final float theta;
-
-    /**
-     * Point of the cube in the form (theta, y) relative to center of tree base
-     */
-
-    public float transformedY;
-    public float transformedTheta;
-    public Vec2D transformedCylinderPoint;
-    public CubeConfig config = null;
+  public final int pixels;
+  public CubeConfig config = null;
 
     Cube(Vec3D globalPosition, Vec3D treePosition, CubeConfig config) {
-        super(Arrays.asList(new LXPoint[] { new LXPoint(globalPosition.x, globalPosition.y, globalPosition.z) }));
-        this.index = this.points.get(0).index;
-        this.size = CUBE_SIZES[config.cubeSizeIndex];
-        this.pixels = PIXELS_PER_CUBE[config.cubeSizeIndex];
-        this.rx = 0;
-        this.ry = 0;
-        this.rz = 0;
-        this.lx = 0;
-        this.ly = 0;
-        this.lz = 0;
-        this.x = globalPosition.x;
-        this.y = globalPosition.y;
-        this.z = globalPosition.z;
-        this.tx = treePosition.x;
-        this.ty = treePosition.y;
-        this.tz = treePosition.z;
-        this.r = (float) Point2D.distance(treePosition.x, treePosition.z, 0, 0);
-        this.theta = 180 + 180 / Utils.PI * Utils.atan2(treePosition.z, treePosition.x);
+      super(globalPosition, treePosition);
+    this.size = CUBE_SIZES[config.cubeSizeIndex];
+    this.pixels = PIXELS_PER_CUBE[config.cubeSizeIndex];
         this.config = config;
     }
-
-    void resetTransform() {
-
-        transformedTheta = theta;
-        transformedY = y;
-    }
-
-    void didTransform() {
-        transformedCylinderPoint = new Vec2D(transformedTheta, transformedY);
-    }
 }
-
 
 
 abstract class Layer extends LXLayer {
@@ -602,7 +512,8 @@ class ModelTransformTask implements LXLoopTask {
 //        model.runShrubTransforms();
     }
 }
-class Geometry{
+
+class Geometry {
     final static int INCHES = 1;
     final static int FEET = 12 * INCHES;
 }
